+++ lib/active_record_compose/attributes.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative "attributes/delegatable_attribute"
 require_relative "attributes/delegation"
 require_relative "attributes/querying"
 
@@ -52,19 +53,26 @@ module ActiveRecordCompose
   #
   module Attributes
     extend ActiveSupport::Concern
-    include ActiveModel::Attributes
 
     included do
+      include ActiveModel::Attributes
       include Querying
 
       # @type self: Class
       class_attribute :delegated_attributes, instance_writer: false
+
+      # steep:ignore:start
+
+      # Returns a array of attribute name.
+      # Attributes declared with `delegate_attribute` are also merged.
+      #
+      # @return [Array<String>] array of attribute name.
+      def self.attribute_names = super + delegated_attributes.to_a.map { _1.attribute_name }
+
+      # steep:ignore:end
     end
 
     module ClassMethods
-      ALLOW_NIL_DEFAULT = Object.new.freeze # steep:ignore
-      private_constant :ALLOW_NIL_DEFAULT
-
       # Defines the reader and writer for the specified attribute.
       #
       # @example
@@ -97,83 +105,38 @@ module ActiveRecordCompose
       #   registration.attributes
       #   # => { "original_attribute" => "qux", "name" => "bar" }
       #
-      def delegate_attribute(*attributes, to:, allow_nil: ALLOW_NIL_DEFAULT) # steep:ignore
-        # steep:ignore:start
+      def delegate_attribute(*attributes, to:, allow_nil: false)
         if to.start_with?("@")
-          suggested_reader_name = to.to_s.sub(/^@+/, "")
-          suggested_method =
-            if to.start_with?("@@")
-              "def #{suggested_reader_name} = #{to}"
-            else
-              "attr_reader :#{suggested_reader_name}"
-            end
-
-          message = <<~MSG
-            Direct use of instance or class variables in `to:` will be removed in the next minor version.
-            Please define a reader method (private is fine) and refer to it by name instead.
-
-            For example,
-                delegate_attribute #{attributes.map { ":#{_1}" }.join(", ")}, to: :#{to}#{", allow_nil: #{allow_nil}" if allow_nil != ALLOW_NIL_DEFAULT}
-
-            Instead of the above, use the following
-                delegate_attribute #{attributes.map { ":#{_1}" }.join(", ")}, to: :#{suggested_reader_name}#{", allow_nil: #{allow_nil}" if allow_nil != ALLOW_NIL_DEFAULT}
-                private
-                #{suggested_method}
-
-          MSG
-          (ActiveRecord.respond_to?(:deprecator) ? ActiveRecord.deprecator : ActiveSupport::Deprecation).warn(message)
+          raise ArgumentError, "Instance variables cannot be specified in delegate to. (#{to})"
         end
-        allow_nil = false if allow_nil == ALLOW_NIL_DEFAULT
-        # steep:ignore:end
 
         delegations = attributes.map { Delegation.new(attribute: _1, to:, allow_nil:) }
         delegations.each { _1.define_delegated_attribute(self) }
 
         self.delegated_attributes = (delegated_attributes.to_a + delegations).reverse.uniq { _1.attribute }.reverse
       end
-
-      # Returns a array of attribute name.
-      # Attributes declared with `delegate_attribute` are also merged.
-      #
-      # @return [Array<String>] array of attribute name.
-      def attribute_names = super + delegated_attributes.to_a.map { _1.attribute_name }
     end
 
-    # Returns a array of attribute name.
-    # Attributes declared with `delegate_attribute` are also merged.
-    #
-    # @return [Array<String>] array of attribute name.
-    def attribute_names = super + delegated_attributes.to_a.map { _1.attribute_name }
+    # steep:ignore:start
+    private_module = Module.new do
+      refine Attributes do
+        private
 
-    # Returns a hash with the attribute name as key and the attribute value as value.
-    # Attributes declared with `delegate_attribute` are also merged.
-    #
-    # @return [Hash] hash with the attribute name as key and the attribute value as value.
-    # @example
-    #   class AccountRegistration < ActiveRecordCompose::Model
-    #     def initialize(account, attributes = {})
-    #       @account = account
-    #       super(attributes)
-    #       models.push(account)
-    #     end
-    #
-    #     attribute :original_attribute, :string, default: "qux"
-    #     delegate_attribute :name, to: :account
-    #
-    #     private
-    #
-    #     attr_reader :account
-    #   end
-    #
-    #   account = Account.new
-    #   account.name = "foo"
-    #
-    #   registration = AccountRegistration.new(account)
-    #
-    #   registration.attributes # => { "original_attribute" => "qux", "name" => "bar" }
-    #
-    def attributes
-      super.merge(*delegated_attributes.to_a.map { _1.attribute_hash(self) })
+        def _merge_delegated_attributes
+          delegated_attributes.to_a.each do |delegation|
+            @attributes[delegation.attribute_name] = DelegatableAttribute.new(delegation, self)
+          end
+        end
+      end
     end
+    using private_module
+    # steep:ignore:end
+
+    # steep:ignore:start
+    def initialize(...)
+      _merge_delegated_attributes
+      super
+    end
+    # steep:ignore:end
   end
 end