+++ lib/active_record_compose/attributes/delegation.rb
@@ -23,9 +23,19 @@ module ActiveRecordCompose
       # @return [String] The attribute name as string
       def attribute_name = attribute.to_s
 
-      # @return [Hash<String, Object>]
-      def attribute_hash(model)
-        { attribute_name => model.public_send(attribute) }
+      def read_attribute(owner)
+        target = owner.send(to)
+        return nil if target.nil? && allow_nil
+
+        target.public_send(reader)
+      end
+
+      def write_attribute(owner, value)
+        target = owner.send(to)
+        return value if target.nil? && allow_nil
+
+        target.public_send(writer, value)
+        value
       end
 
       private